move pdp.py script contents to module directory
This is a more robust structure for future creation of multiple entrypoints,
and implementing testing via nosetests
